Add files via upload
About this Project: 
The Task Management System using PHP and MySQL can help a particular company manage its project task progress. The system has three system users: the Admin, Project Manager, and the Regular Employee. Admin users can only create the Task Management System users. The admin user or the Project Managers will create a new project first and some essential details and references of the users. When creating a project, the admin or project managers must list all the employees to handle the project’s tasks. After completing the project, the regular employees are limited only to adding their work progress for the project they are assigned. They are also allowed to edit the data of the progress that they only submitted. Then, as the employees regularly update the system about their progress, the project managers will read or scan their activities. The project manager can check or decide if a particular task is done and need to update the system’s task status. For the system’s printable report, only the admin users and project managers have access to this function or feature of the system.